+++ src/components/admin/MpAdminCarousel.vue
@@ -2,18 +2,18 @@
 import { defineAsyncComponent, onMounted, ref } from 'vue'
 import { useConfirm } from 'primevue/useconfirm'
 import { useToast } from 'primevue/usetoast'
-import { useCatalogsStore} from '@/store/catalogs.js'
+import { useCarouselStore } from '@/store/carousel.js'
 
 const MpAdminModal = defineAsyncComponent(/* webpackChunkName: "mpAdminModal" */() => import('@/components/admin/MpAdminModal.vue'))
 
-const catalogsStore = useCatalogsStore()
+const carouselStore = useCarouselStore()
 const confirm = useConfirm()
 const toast = useToast()
 
-const openModal = ref(false)
+const openModalCarousel = ref(false)
 
 const manageValueModal = (value) => {
-  openModal.value = value
+  openModalCarousel.value = value
 }
 
 const configModal = ref({})
@@ -21,8 +21,8 @@ const configModal = ref({})
 const manageCatalog = (action, type, data = null) => {
   if (action === 'add') {
     configModal.value = {
-      header: 'Agregar catálogo',
-      description: 'Completa los campos para agregar un nuevo catálogo.',
+      header: 'Agregar imagen en el carrusel',
+      description: 'Completa los campos para agregar una nueva imagen en el carusel.',
       acceptButton: 'Agregar',
       cancelButton: 'Cancelar',
       type,
@@ -33,8 +33,8 @@ const manageCatalog = (action, type, data = null) => {
 
   if (action === 'edit') {
     configModal.value = {
-      header: 'Editar catálogo',
-      description: 'Completa los campos para editar el catálogo.',
+      header: 'Editar imagen del carrusel',
+      description: 'Completa los campos para editar la imagen del carrusel.',
       acceptButton: 'Editar',
       cancelButton: 'Cancelar',
       type,
@@ -42,12 +42,12 @@ const manageCatalog = (action, type, data = null) => {
       data
     }
   }
-  openModal.value = true
+  openModalCarousel.value = true
 }
 
-const confirmDelete = (id, image) => {
+const confirmDeleteCarousel = (id, image) => {
   confirm.require({
-    message: '¿Quieres eliminar este catálogo?',
+    message: '¿Quieres eliminar esta imagen del carrusel?',
     header: 'Confirmar',
     icon: 'pi pi-info-circle mr-2',
     rejectLabel: 'Cancelar',
@@ -55,8 +55,8 @@ const confirmDelete = (id, image) => {
     rejectClass: 'transparent',
     acceptClass: 'bg-danger text-white',
     accept: () => {
-      toast.add({ severity: 'info', summary: 'Eliminado', detail: 'El catálogo ha sido eliminado.', life: 3000 })
-      catalogsStore.deleteCatalog(id, image)
+      toast.add({ severity: 'info', summary: 'Eliminado', detail: 'La imagen del carrusel ha sido eliminada.', life: 3000 })
+      carouselStore.deleteCarousel(id, image)
     },
     reject: () => {
       toast.add({ severity: 'error', summary: 'Cancelado', detail: 'Se ha cancelado la operación.', life: 3000 })
@@ -65,44 +65,44 @@ const confirmDelete = (id, image) => {
 }
 
 onMounted(() => {
-  catalogsStore.getCatalogs()
+  carouselStore.getCarousel()
 })
 </script>
 
 <template>
   <MpAdminModal
-    :visible="openModal"
+    :visible="openModalCarousel"
     :configModal="configModal"
     @manageModal="manageValueModal"
   />
-  <div v-if="catalogsStore.isLoading" class="container mx-auto grid gap-4 mt-1">
+  <div v-if="carouselStore.isLoading" class="container mx-auto grid gap-4 mt-1">
     <div class="flex flex-wrap align-items-center justify-content-between gap-2">
-      <span class="text-xl text-900 font-bold">Catálogos</span>
+      <span class="text-xl text-900 font-bold">Carrusel de imágenes</span>
     </div>
     <template v-for="loader in 12">
       <Skeleton height="2rem" class="mb-2">{{loader}}</Skeleton>
     </template>
   </div>
   <template v-else>
-    <DataTable :value="catalogsStore.catalogs" tableStyle="min-width: 50rem">
+    <DataTable :value="carouselStore.carousel" tableStyle="min-width: 50rem">
       <template #header>
         <div class="flex flex-wrap align-items-center justify-content-between gap-2">
-          <span class="text-xl text-900 font-bold">Catálogos</span>
+          <span class="text-xl text-900 font-bold">Carrusel de imágenes</span>
           <Button
             icon="pi pi-refresh"
             rounded
             raised
-            v-tooltip.top="`Actualizar catálogos`"
-            @click="catalogsStore.getCatalogs(true)"
-            :disabled="catalogsStore.catalogs.length === 0"
+            v-tooltip.top="`Actualizar imágenes en el carrusel`"
+            @click="carouselStore.getCarousel(true)"
+            :disabled="carouselStore.carousel.length === 0"
           />
           <Button
             icon="pi pi-plus"
             rounded
             raised
-            v-tooltip.top="`Agregar catálogo`"
+            v-tooltip.top="`Agregar imagen en el carrusel`"
             severity="info"
-            @click="manageCatalog('add', 'catalog')"
+            @click="manageCatalog('add', 'carousel')"
           />
         </div>
       </template>
@@ -112,13 +112,18 @@ onMounted(() => {
           <img
             :src="slotProps.data.image"
             :alt="slotProps.data.title"
-            class="w-6rem border-round w-20"
+            class="w-6rem border-round w-[20rem]"
           />
         </template>
       </Column>
       <Column header="Link">
         <template #body="slotProps">
-          <a :href="slotProps.data.link" target="_blank" class="text-blue-500 hover:text-blue-700">
+          <a
+            v-if="slotProps.data.link"
+            :href="slotProps.data.link"
+            target="_blank"
+            class="text-blue-500 hover:text-blue-700"
+          >
             {{ slotProps.data.link }} <span class="pi pi-external-link"></span>
           </a>
         </template>
@@ -129,18 +134,18 @@ onMounted(() => {
             <Button
               icon="pi pi-pencil"
               class="p-button-rounded p-button-success p-button-outlined"
-              v-tooltip.top="`Editar catálogo`"
+              v-tooltip.top="`Editar imagen del carrusel`"
               severity="info" rounded outlined
-              aria-label="Editar catálogo"
-              @click="manageCatalog('edit', 'catalog', slotProps.data)"
+              aria-label="Editar imagen del carrusel"
+              @click="manageCatalog('edit', 'carousel', slotProps.data)"
             />
             <Button
               icon="pi pi-trash"
               class="p-button-rounded p-button-danger p-button-outlined"
-              v-tooltip.top="`Eliminar catálogo`"
+              v-tooltip.top="`Eliminar imagen del carrusel`"
               severity="danger" text rounded
-              aria-label="Eliminar catálogo"
-              @click="confirmDelete(slotProps.data.id, slotProps.data.image)"
+              aria-label="Eliminar imagen del carrusel"
+              @click="confirmDeleteCarousel(slotProps.data.id, slotProps.data.image)"
             />
           </div>
         </template>