+++ src/store/carousel.js
@@ -1,6 +1,7 @@
 import { defineStore } from 'pinia'
 import { db } from '../../firebase.js'
-import { collection, getDocs } from 'firebase/firestore'
+import { addDoc, collection, deleteDoc, doc, getDocs, updateDoc } from 'firebase/firestore'
+import { getDownloadURL, getStorage, ref as storageRef, uploadBytes, deleteObject } from 'firebase/storage'
 import { daysDifferenceFromMidnight } from '../utils/index.js'
 
 export const useCarouselStore = defineStore('carousel', {
@@ -9,18 +10,20 @@ export const useCarouselStore = defineStore('carousel', {
         isLoading: true
     }),
     actions: {
-        async getCarousel() {
+        async getCarousel(update = false) {
             this.isLoading = true
-            const carouselLocalStorage = localStorage.getItem('carousel')
-            const lastUpdateLocalStorage = localStorage.getItem('carouselLastUpdate')
-            if (carouselLocalStorage && lastUpdateLocalStorage) {
-                const lastUpdateDate = new Date(lastUpdateLocalStorage)
-                const now = new Date()
-                const diffInDays = daysDifferenceFromMidnight(lastUpdateDate, now);
-                if (diffInDays < 1) {
-                    this.carousel = JSON.parse(carouselLocalStorage)
-                    this.isLoading = false
-                    return
+            if (!update) {
+                const carouselLocalStorage = localStorage.getItem('carousel')
+                const lastUpdateLocalStorage = localStorage.getItem('carouselLastUpdate')
+                if (carouselLocalStorage && lastUpdateLocalStorage) {
+                    const lastUpdateDate = new Date(lastUpdateLocalStorage)
+                    const now = new Date()
+                    const diffInDays = daysDifferenceFromMidnight(lastUpdateDate, now);
+                    if (diffInDays < 1) {
+                        this.carousel = JSON.parse(carouselLocalStorage)
+                        this.isLoading = false
+                        return
+                    }
                 }
             }
 
@@ -35,6 +38,50 @@ export const useCarouselStore = defineStore('carousel', {
             localStorage.setItem('carousel', JSON.stringify(items))
             localStorage.setItem('carouselLastUpdate', new Date().toISOString())
             this.isLoading = false
+        },
+        
+        async deleteCarousel(id, imageUrl) {
+            this.isLoading = true
+            await deleteDoc(doc(db, 'carousel', id))
+            if (imageUrl) {
+                const storage = getStorage()
+                const imageRef = storageRef(storage, imageUrl)
+                try {
+                    await deleteObject(imageRef)
+                } catch (error) {
+                    console.error("Error al eliminar la imagen:", error)
+                }
+            }
+            await this.getCarousel(true)
+        },
+        
+        async _uploadImageCarousel (file) {
+            console.log({file})
+            if (!file) return
+            
+            const storage = getStorage();
+            const fileRef = storageRef(storage, `carousel/${file.name}`)
+            await uploadBytes(fileRef, file)
+            return await getDownloadURL(fileRef)
+        },
+        
+        async addImageCarousel(data) {
+            console.log({data})
+            this.isLoading = true
+            const image = await this._uploadImageCarousel(data.image)
+            await addDoc(collection(db, 'carousel'), { ...data, image })
+            await this.getCarousel(true)
+        },
+        
+        async editImageCarousel(data) {
+            this.isLoading = true
+            let image = data.image
+            if (data.image instanceof File) {
+                image = await this._uploadImageCarousel(data.image)
+            }
+            await doc(db, 'carousel', data.id)
+            await updateDoc(doc(db, 'carousel', data.id), { ...data, image })
+            await this.getCarousel(true)
         }
     }
 })